fix: fix mismatched parsing data types

+++ src/models/earning_history_model.rs
@@ -26,12 +26,12 @@ pub struct PoolEarningSummary{
 pub struct PoolEarningHistory{
     pub _id : ObjectId,
     pub pool : String,
-    pub asset_liquidity_fees : u64,
+    pub asset_liquidity_fees : f64,
     pub earning : u64,
-    pub rewards : u64,
-    pub rune_liquidity_fees : u64,
-    pub saver_earning : u64,
-    pub total_liquidity_fees_rune : u64,
+    pub rewards : f64,
+    pub rune_liquidity_fees : f64,
+    pub saver_earning : f64,
+    pub total_liquidity_fees_rune : f64,
     pub start_time : i64,
     pub end_time : i64,
     pub earnings_summary : ObjectId
